resolve first correction
"Section 2.1: Provide, and refer to, an illustrative example of a very simple variation graph,
showing explicitly the paths, nodes, edges, as constructed in one or two ways based e.g. on a
minimal alignment of reads from a diploid genome or short sequences."

This is resolved by a new figure which is referenced in the text.