improve the introduction to explain the structure of the first chapter
In my viva I understood that the purpose of the initial section in the first
chapter was not clarified. Here, I attempt to do so in the final paragraphs of
the introductory section, and provide advice to the reader about what sections
to focus on depending on their prior knowledge.